+++ tests/unit-tests/VolleyM.Domain.Players.UnitTests/IPlayersTestFixture.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 using VolleyM.Domain.Contracts;
+using VolleyM.Domain.Players.Handlers;
 using VolleyM.Domain.Players.PlayerAggregate;
 using VolleyM.Domain.Players.UnitTests.Fixture;
 using VolleyM.Domain.UnitTests.Framework;
@@ -18,5 +19,7 @@ namespace VolleyM.Domain.Players.UnitTests
 		void MockNextRandomId(string id);
 
 		Task VerifyPlayerNotCreated(Player expectedPlayer);
+
+		void SetupPlayerName(IPlayerNameRequest request);
 	}
 }
